feat. 예외처리 추가, controller return 구문 변경
